[DEMO-2561] Deduplicate common methods in lambda, tron entrypoint
* Reducing code duplication by moving the get_inventory_data &
selector methods into a single location, which both entrypoints refer
to.
* Also cleaned up imports.
